Separated rooms from meetings
