+++ spec/controllers/meetings_controller_spec.rb
@@ -1,0 +1,78 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Api::V1::MeetingsController, type: :controller do
+  let(:user) { create(:user) }
+
+  before do
+    request.headers['ACCEPT'] = 'application/json'
+    session[:user_id] = user.id
+  end
+
+  describe '#start' do
+    let(:room) { create(:room, user:) }
+
+    before do
+      allow_any_instance_of(BigBlueButtonApi)
+        .to receive(:start_meeting)
+        .and_return(true)
+    end
+
+    it 'makes a call to the MeetingStarter service with the right values' do
+      logout = 'http://example.com'
+      request.env['HTTP_REFERER'] = logout
+
+      expect(MeetingStarter).to receive(:new).with(room:, logout_url: logout)
+
+      post :start, params: { friendly_id: room.friendly_id }
+    end
+
+    it 'makes a call to the BigBlueButtonApi to get the join url' do
+      expect_any_instance_of(BigBlueButtonApi)
+        .to receive(:join_meeting)
+        .with(room:, name: user.name, role: 'Moderator')
+
+      post :start, params: { friendly_id: room.friendly_id }
+    end
+
+    it 'returns the join url to the front-end for redirecting' do
+      allow_any_instance_of(BigBlueButtonApi)
+        .to receive(:join_meeting)
+        .and_return('https://example.com')
+
+      post :start, params: { friendly_id: room.friendly_id }
+
+      expect(response).to have_http_status(:created)
+      expect(JSON.parse(response.body)['data']['join_url']).to eq('https://example.com')
+    end
+  end
+
+  describe '#join' do
+    it 'calls the BigBlueButton service with the right values' do
+      room = create(:room, user:)
+
+      expect_any_instance_of(BigBlueButtonApi).to receive(:join_meeting).with(room:, name: user.name, role: 'Viewer')
+      get :join, params: { friendly_id: room.friendly_id, name: user.name }
+    end
+  end
+
+  describe '#status' do
+    it 'calls the BigBlueButton service with the right values' do
+      room = create(:room, user:)
+
+      expect_any_instance_of(BigBlueButtonApi).to receive(:meeting_running?).with(room:)
+
+      get :status, params: { friendly_id: room.friendly_id, name: user.name }
+    end
+
+    it 'gets the joinUrl if the meeting is running' do
+      room = create(:room, user:)
+
+      allow_any_instance_of(BigBlueButtonApi).to receive(:meeting_running?).and_return(true)
+      expect_any_instance_of(BigBlueButtonApi).to receive(:join_meeting).with(room:, name: user.name, role: 'Viewer')
+
+      get :status, params: { friendly_id: room.friendly_id, name: user.name }
+    end
+  end
+end
